delete base image manager
