+++ image/image.go
@@ -2,16 +2,12 @@ package image
 
 // NewImageService return the image service
 func NewImageService() Service {
-	return DefaultImageService{
-		BaseImageManager{},
-	}
+	return DefaultImageService{}
 }
 
 // NewImageMetadataService return the MetadataService
 func NewImageMetadataService() MetadataService {
-	return DefaultImageMetadataService{
-		BaseImageManager{},
-	}
+	return DefaultImageMetadataService{}
 }
 
 // NewImageFileService return the file Service
